# ARROW-6381: [C++] BufferOutputStream::Write does extra work that slows down small writes