Before:

```
-------------------------------------------------------------------------------
Benchmark                                        Time           CPU Iterations
-------------------------------------------------------------------------------
BufferOutputStreamSmallWrites/real_time  218998806 ns  218938001 ns         18   584.478MB/s
```

After

```
-------------------------------------------------------------------------------
Benchmark                                        Time           CPU Iterations
-------------------------------------------------------------------------------
BufferOutputStreamSmallWrites/real_time   11439597 ns   11436629 ns         61   2.73169GB/s
```