@wesm 2019-08-28
I found this when looking at Parquet perf data (this is used there, it might be better to use BufferBuilder instead)