@pitrou 2019-08-29
Interesting. I wonder if `BufferOutputStream` should rely on `BufferBuilder` or if that would degrade performance.