@pitrou 2019-08-29
Current benchmarks here:
```
-------------------------------------------------------------------------------
Benchmark                                        Time           CPU Iterations
-------------------------------------------------------------------------------
BufferOutputStreamTinyWrites/real_time    40346155 ns   40231527 ns         52   793.136MB/s
BufferOutputStreamSmallWrites/real_time   15462451 ns   15438635 ns        100   2.02099GB/s
BufferOutputStreamLargeWrites/real_time   14104049 ns   14066428 ns        178   2.17907GB/s
```