@pitrou 2019-08-29
For the record, here are the results on a Windows 7 VM. It will be interesting to check if mimalloc improves those numbers:
```
--------------------------------------------------------------------------------
------------------
Benchmark                                        Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------
------------------
BufferOutputStreamTinyWrites/real_time    90775594 ns     90725806 ns 31 bytes_per_second=352.518M/s
BufferOutputStreamSmallWrites/real_time   32634961 ns     32567771 ns 83 bytes_per_second=980.525M/s
BufferOutputStreamLargeWrites/real_time   31290053 ns     31250000 ns 91 bytes_per_second=1005.79M/s
```