@wesm 2019-08-29
Thanks @pitrou!

> Interesting. I wonder if BufferOutputStream should rely on BufferBuilder or if that would degrade performance.

Probably, it seems a bit weird to have two implementations of nearly the same thing. https://issues.apache.org/jira/browse/ARROW-6388